refactor: confirm delete prompt version
