+++ client/src/components/Prompts/DeleteVersion.tsx
@@ -1,0 +1,60 @@
+import { Button, Dialog, DialogTrigger, Label } from '~/components/ui';
+import DialogTemplate from '~/components/ui/DialogTemplate';
+import { TrashIcon } from '~/components/svg';
+import { useLocalize } from '~/hooks';
+
+const DeleteVersion = ({
+  name,
+  disabled,
+  selectHandler,
+}: {
+  name: string;
+  disabled?: boolean;
+  selectHandler: () => void;
+}) => {
+  const localize = useLocalize();
+
+  return (
+    <Dialog>
+      <DialogTrigger asChild>
+        <Button
+          size={'sm'}
+          className="h-10 w-10 border border-transparent bg-red-600 text-red-500 transition-all hover:bg-red-700 dark:bg-red-600 dark:hover:bg-red-800"
+          disabled={disabled}
+          onClick={(e) => {
+            e.stopPropagation();
+          }}
+        >
+          <TrashIcon className="icon-lg cursor-pointer text-white dark:text-white" />
+        </Button>
+      </DialogTrigger>
+      <DialogTemplate
+        showCloseButton={false}
+        title={localize('com_ui_delete_prompt')}
+        className="max-w-[450px]"
+        main={
+          <>
+            <div className="flex w-full flex-col items-center gap-2">
+              <div className="grid w-full items-center gap-2">
+                <Label
+                  htmlFor="dialog-delete-confirm-prompt"
+                  className="text-left text-sm font-medium"
+                >
+                  {localize('com_ui_delete_confirm_prompt_version_var', name)}
+                </Label>
+              </div>
+            </div>
+          </>
+        }
+        selection={{
+          selectHandler,
+          selectClasses:
+            'bg-red-700 dark:bg-red-600 hover:bg-red-800 dark:hover:bg-red-800 text-white',
+          selectText: localize('com_ui_delete'),
+        }}
+      />
+    </Dialog>
+  );
+};
+
+export default DeleteVersion;
