+++ client/src/localization/languages/Eng.ts
@@ -258,6 +258,8 @@ export default {
   com_ui_shared_link_not_found: 'Shared link not found',
   com_ui_delete_conversation: 'Delete chat?',
   com_ui_delete_confirm: 'This will delete',
+  com_ui_delete_confirm_prompt_version_var:
+    'This will delete the selected version for "{0}." If no other versions exist, the prompt will be deleted.',
   com_ui_delete_assistant_confirm:
     'Are you sure you want to delete this Assistant? This cannot be undone.',
   com_ui_rename: 'Rename',
